+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - expand as query parameter to pix key resource
+- reason parameter on pix key cancellation resource
 
 ## [0.18.0] - 2025-04-23
 ### Added